+++ packages/orq-rc/src/orq_ai_sdk/models/evaluatorresponsellm.py
@@ -250,9 +250,9 @@ class EvaluatorResponseLlm(BaseModel):
 
     mode: Mode
 
-    created: Optional[str] = "2026-07-02T12:43:37.944Z"
+    created: Optional[str] = "2026-07-02T16:16:01.733Z"
 
-    updated: Optional[str] = "2026-07-02T12:43:37.944Z"
+    updated: Optional[str] = "2026-07-02T16:16:01.734Z"
 
     updated_by_id: OptionalNullable[str] = UNSET