+++ packages/orq-rc/src/orq_ai_sdk/models/retrievedatasourceop.py
@@ -90,7 +90,7 @@ class RetrieveDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KWHDQ8V3FT3K00HH0A5DQS5C"
+        "01KWHSVZPA2KGNEEG115Z8JPPE"
     )
     r"""The unique identifier of the data source"""