+++ packages/orq-rc/src/orq_ai_sdk/models/updatedatasetop.py
@@ -160,7 +160,7 @@ class UpdateDatasetResponseBody(BaseModel):
     created: Optional[datetime] = None
     r"""The date and time the resource was created"""
 
-    updated: Optional[datetime] = parse_datetime("2026-07-02T12:43:35.190Z")
+    updated: Optional[datetime] = parse_datetime("2026-07-02T16:15:59.901Z")
     r"""The date and time the resource was last updated"""
 
     @model_serializer(mode="wrap")